backend/requirements.txt
Flask==2.3.2
flask_cors==4.0.0
Pillow==10.0.1
numpy==1.26.0
